Блокировка изменения в диалоге закрытия
