+++ Source/Applications/Desktop/Vodovoz/Dialogs/Logistic/RouteListClosingDlg.RouteListActions.cs
@@ -1,0 +1,25 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz
+{
+	public partial class RouteListClosingDlg
+	{
+		enum RouteListActions
+		{
+			[Display(Name = "Новый штраф")]
+			CreateNewFine,
+			[Display(Name = "Перенести разгрузку в другой МЛ")]
+			TransferReceptionToAnotherRL,
+			[Display(Name = "Перенести разгрузку в этот МЛ")]
+			TransferReceptionToThisRL,
+			[Display(Name = "Перенести адреса в этот МЛ")]
+			TransferAddressesToThisRL,
+			[Display(Name = "Перенести адреса из этого МЛ")]
+			TransferAddressesToAnotherRL
+
+		}
+
+		#endregion
+	}
+
+}
