+++ Source/Applications/Desktop/Vodovoz/Dialogs/Logistic/RouteListClosingDlg.RouteListPrintDocuments.cs
@@ -1,0 +1,20 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz
+{
+	public partial class RouteListClosingDlg
+	{
+		public enum RouteListPrintDocuments
+		{
+			[Display(Name = "Все")]
+			All,
+			[Display(Name = "Маршрутный лист")]
+			RouteList,
+			[Display(Name = "Штрафы")]
+			Fines
+		}
+
+		#endregion
+	}
+
+}
